1st version of NRPPa Position Information Transfer Procedure

+++ openair3/NGAP/ngap_gNB_itti_messaging.h
@@ -40,7 +40,7 @@ void ngap_gNB_itti_send_nas_downlink_ind(instance_t instance, uint32_t gNB_ue_ng
 
 void ngap_gNB_itti_send_sctp_close_association(instance_t instance, sctp_assoc_t assoc_id);
 
-void ngap_gNB_itti_send_DownlinkUEAssociatedNRPPa(instance_t instance, uint32_t gNB_ue_ngap_id, uint32_t amf_ue_ngap_id,  uint8_t *routingId_buffer, uint32_t routingId_buffer_length, uint8_t *nrppa_pdu, uint32_t nrppa_pdu_length);
+void ngap_gNB_itti_send_DownlinkUEAssociatedNRPPa(instance_t instance, uint32_t gNB_ue_ngap_id, uint64_t amf_ue_ngap_id,  uint8_t *routingId_buffer, uint32_t routingId_buffer_length, uint8_t *nrppa_pdu, uint32_t nrppa_pdu_length);
 
 void ngap_gNB_itti_send_DownlinkNonUEAssociatedNRPPa(instance_t instance, uint8_t *routingId_buffer, uint32_t routingId_buffer_length, uint8_t *nrppa_pdu, uint32_t nrppa_pdu_length);
